outputs: Toggle between each output, assigning each new surface
This does a fancy dance between outputs on the v128 to ensure that windows are
distributed evenly across each display as surfaces come and go. This is a pretty
naiive implementation, though, so if a window appears, disappears, and
reappears, it will end up on the other output.

For now, this gets us to our goal of displaying x128 on two heads.